using-Boost-ASIO_instead

+++ core/core.h
@@ -2,7 +2,7 @@
 #define _Included_Core_
 
 namespace core {
-	void startMining(const char*, const unsigned int, const char*, const char*);
+	void startMining(const char*, const unsigned short, const char*, const char*);
 	void stopMining();
 }
