+++ core/pass_function_set.h
@@ -7,7 +7,7 @@ namespace function_set {
 	extern void (*afterStop) ();
 	//socket connection
 	//return false cause error or has connection 
-	extern void (*openConnection) (const char*,const unsigned int);
+	extern void (*openConnection) (const char*,const unsigned short);
 	extern const char*(*recvConnection) ();
 	extern bool (*sendMessage) (const char *msg);
 	//return false cause error or no connection
